FIX src compileoptions -- auto-detect UNIX-ENVIRONMENT
'DONE' did not work on building the system in Windows if the source
files have LF line endings and UNIX-ENVIRONMENT was not set to TRUE.

Now the appropriate value for UNIX-ENVIRONMENT option (that determinates
line endings mode) is auto-detected for the case if it is absent in
compile.ini.

We should set UNIX-LINES mode in the target system if the source files
have LF line endings.