FIX src tc -- bug in the target compiler `postpone`
When the argument of `postpone` is an immediate word,
its address is stored by `lit,`, and it was not
converted to the virtual address space before storing.

The bug is effective only on Linux, where the virtual
address space is used, and when `postpone` is used in
the target code (there were no such cases before this
commit, but there will be soon).